Add styling to root element top prevent skipping bug in some recent Blink versions
The 'overflow-anchor' CSS style implementation in recent versions of
Chrome, Opera and Android browser contains a bug which causes the root
element to fire excess scroll events with different scrollTop property
values. This causes a endless loop of render calls until the bottom of
the list is reached.

As 'overflow-anchor' doesn't have any effect on the root element, this
is disabled as a workaround for the bug.